Refactor: Update namespace
- src/Frouros/Net/Endianness.cs: Update namespace
- src/Frouros/Net/PacketLog.cs: Update namespace
