+++ src/Frouros/Net/PacketLog.cs
@@ -1,24 +1,23 @@
-//    Copyright 2023 Yeong-won Seo
+//     Copyright 2023 Yeong-won Seo
 // 
-//    Licensed under the Apache License, Version 2.0 (the "License");
-//    you may not use this file except in compliance with the License.
-//    You may obtain a copy of the License at
+//     Licensed under the Apache License, Version 2.0 (the "License");
+//     you may not use this file except in compliance with the License.
+//     You may obtain a copy of the License at
 // 
-//        http://www.apache.org/licenses/LICENSE-2.0
+//         http://www.apache.org/licenses/LICENSE-2.0
 // 
-//    Unless required by applicable law or agreed to in writing, software
-//    distributed under the License is distributed on an "AS IS" BASIS,
-//    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//    See the License for the specific language governing permissions and
-//    limitations under the License.
+//     Unless required by applicable law or agreed to in writing, software
+//     distributed under the License is distributed on an "AS IS" BASIS,
+//     WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//     See the License for the specific language governing permissions and
+//     limitations under the License.
 
-using System.Buffers.Binary;
 using System.Net;
 using System.Net.Sockets;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
-namespace Frouros;
+namespace Frouros.Net;
 
 internal partial struct NativePacketLog
 {
@@ -100,7 +99,7 @@ public readonly struct PacketLog(
     LxProto      lx,
     IPEndPoint   source,
     IPEndPoint   destination,
-    uint size)
+    uint         size)
 {
     public readonly DateTime     Timestamp   = timestamp;
     public readonly ProtocolType L3          = l3;
@@ -113,15 +112,15 @@ public readonly struct PacketLog(
     {
         UInt128 src = 0;
         UInt128 dst = 0;
-        
+
         unsafe
         {
             fixed (byte* pSrc = Source.Address.GetAddressBytes())
                 Unsafe.CopyBlock(&src, pSrc, 16);
             fixed (byte* pDst = Source.Address.GetAddressBytes())
-                Unsafe.CopyBlock(&dst, pDst, 16);   
+                Unsafe.CopyBlock(&dst, pDst, 16);
         }
-        
+
         var ts = (ulong)((DateTimeOffset)Timestamp).ToUnixTimeSeconds();
         var l2 = NativePacketLog.GetL2Family(Source.AddressFamily);
         var l3 = (byte)L3;
